+++ src/app/_constant/modal.ts
@@ -2,3 +2,9 @@ export const MODAL = {
   POST_FEED: "새 게시물 만들기",
   TEST: "TEST MODAL",
 };
+
+export const POST_MODAL = {
+  UPLOAD: "img upload",
+  PREVIEW: "img preview & edit",
+  WRITE: "write content",
+};
